Moved common functions for the drivers to a BaseDriver class in the common folder. Fixed notebooks. We still need to updated some of the information in the tutorial to explicit that the integrator is initialised through the driver
