+++ fidimag/atomistic/atomistic_driver.py
@@ -1,16 +1,17 @@
 from __future__ import division
 from __future__ import print_function
 
+from fidimag.common.driver_base import DriverBase
+
 import fidimag.extensions.clib as clib
 import fidimag.extensions.cvode as cvode
-from fidimag.common.save_vtk import SaveVTK
+from fidimag.common.vtk import VTK
 import fidimag.common.constant as const
 
 import numpy as np
-import os
 
 
-class AtomisticDriver(object):
+class AtomisticDriver(DriverBase):
     """
 
     A class with shared methods and properties for different drivers to solve
@@ -40,6 +41,8 @@ class AtomisticDriver(object):
                  data_saver,
                  use_jac
                  ):
+
+        super(AtomisticDriver, self).__init__()
 
         # These are (ideally) references to arrays taken from the Simulation
         # class. Variables with underscore are arrays changed by a property in
@@ -72,9 +75,6 @@ class AtomisticDriver(object):
         self.n_nonzero = self.mesh.n  # number of spins that are not zero
                                       # We check this in the set_Ms function
 
-        # To save VTK files:
-        self.vtk = SaveVTK(self.mesh, name=name)
-
         if use_jac is not True:
             self.integrator = cvode.CvodeSolver(self.spin, self.sundials_rhs)
         else:
@@ -85,15 +85,25 @@ class AtomisticDriver(object):
 
         self.set_tols()
 
-        # When initialising the integrator in the self.integrator call, the CVOde
-        # class calls the set_initial_value function (with flag_m=0), which
-        # initialises a new integrator and allocates memory in this process.
-        # Now, when we set the magnetisation, we will use the same memory
-        # setting this flag_m to 1, so instead of calling CVodeInit we call
-        # CVodeReInit. If don't, memory is allocated in every call of set_m
+        # When initialising the integrator in the self.integrator call, the
+        # CVOde class calls the set_initial_value function (with flag_m=0),
+        # which initialises a new integrator and allocates memory in this
+        # process.  Now, when we set the magnetisation, we will use the same
+        # memory setting this flag_m to 1, so instead of calling CVodeInit we
+        # call CVodeReInit. If don't, memory is allocated in every call of
+        # set_m
         self.flag_m = 1
 
-        # ---------------------------------------------------------------------
+        # Factor for the dmdt magnitude in the relaxation function
+        self._dmdt_factor = 1.
+
+        # Savers --------------------------------------------------------------
+
+        # VTK saver for the magnetisation/spin field
+        self.VTK = VTK(self.mesh,
+                       directory='{}_vtks'.format(self.name),
+                       filename='m'
+                       )
 
         self.data_saver = data_saver
 
@@ -129,46 +139,9 @@ class AtomisticDriver(object):
         self.gamma = gamma
         self.do_precession = True
 
-    def set_tols(self, rtol=1e-8, atol=1e-10):
-        """
-        Set the relative and absolute tolerances for the CVODE integrator
-        """
-        self.integrator.set_options(rtol, atol)
-
     # Don't know the uselfulness of this function:
     # def set_options(self, rtol=1e-8, atol=1e-10):
     #     self.set_tols(rtol, atol)
-
-    def compute_effective_field(self, t):
-        """
-        Compute the effective field from the simulation interactions,
-        calling the method from the micromagnetic Energy class
-        """
-        self.field[:] = 0
-        for obj in self.interactions:
-            self.field += obj.compute_field(t)
-
-    def compute_effective_field_jac(self, t, spin):
-        self.field[:] = 0
-        for obj in self.interactions:
-            if obj.jac:
-                self.field += obj.compute_field(t, spin=spin)
-
-    def compute_dmdt(self, dt):
-        m0 = self.spin_last
-        m1 = self.spin
-        dm = (m1 - m0).reshape((-1, 3))
-        max_dm = np.max(np.sqrt(np.sum(dm**2, axis=1)))
-        max_dmdt = max_dm / dt
-        return max_dmdt
-
-    def stat(self):
-        return self.integrator.stat()
-
-    def reset_integrator(self, t=0):
-        self.integrator.reset(self.spin, t)
-        self.t = t  # also reinitialise the simulation time and step
-        self.step = 0
 
     def sundials_rhs(self, t, y, ydot):
         """
@@ -193,67 +166,6 @@ class AtomisticDriver(object):
 
         return 0
 
-    def run_until(self, t):
-
-        if t <= self.t:
-            if t == self.t and self.t == 0.0:
-                self.compute_effective_field(t)
-                self.data_saver.save()
-            return
-
-        ode = self.integrator
-
-        self.spin_last[:] = self.spin[:]
-
-        flag = ode.run_until(t)
-
-        if flag < 0:
-            raise Exception("Run cython run_until failed!!!")
-
-        self.spin[:] = ode.y[:]
-
-        self.t = t
-        self.step += 1
-
-        # update field before saving data
-        self.compute_effective_field(t)
-        self.data_saver.save()
-
-    def relax(self, dt=1e-11, stopping_dmdt=0.01,
-              max_steps=1000, save_m_steps=100, save_vtk_steps=100):
-
-        for i in range(0, max_steps + 1):
-
-            cvode_dt = self.integrator.get_current_step()
-
-            increment_dt = dt
-
-            if cvode_dt > dt:
-                increment_dt = cvode_dt
-
-            self.run_until(self.t + increment_dt)
-
-            if save_vtk_steps is not None:
-                if i % save_vtk_steps == 0:
-                    self.save_vtk()
-            if save_m_steps is not None:
-                if i % save_m_steps == 0:
-                    self.save_m()
-
-            dmdt = self.compute_dmdt(increment_dt)
-
-            print('step=%d, time=%0.3g, max_dmdt=%0.3g ode_step=%0.3g'
-                  % (self.step, self.t, dmdt, cvode_dt))
-
-            if dmdt < stopping_dmdt:
-                break
-
-        if save_m_steps is not None:
-            self.save_m()
-
-        if save_vtk_steps is not None:
-            self.save_vtk()
-
     # -------------------------------------------------------------------------
     # Save functions ----------------------------------------------------------
     # -------------------------------------------------------------------------
@@ -267,34 +179,10 @@ class AtomisticDriver(object):
         NOTE: It is recommended to use a *cell to point data* filter in
         Paraview or Mayavi to plot the vector field
         """
-        self.vtk.save_vtk(self.spin.reshape(-1, 3),
-                          self._mu_s / const.mu_B,
-                          step=self.step
-                          )
+        self.VTK.reset_data()
 
-    def save_m(self):
-        """
-        Save the magnetisation/spin vector field as a numpy array in
-        a NPY file. The files are saved in the `{name}_npys` folder, where
-        `{name}` is the simulation name, with the file name `m_{step}.npy`
-        where `{step}` is the simulation step (from the integrator)
-        """
-        if not os.path.exists('%s_npys' % self.name):
-            os.makedirs('%s_npys' % self.name)
-        name = '%s_npys/m_%g.npy' % (self.name, self.step)
-        np.save(name, self.spin)
+        # Here we save both Ms and spins as cell data
+        self.VTK.save_scalar(self._mu_s / const.mu_B, name='mu_s')
+        self.VTK.save_vector(self.spin.reshape(-1, 3), name='spins')
 
-    def save_skx(self):
-        """
-        Save the skyrmion number density (sk number per mesh site)
-        as a numpy array in a NPY file.
-        The files are saved in the `{name}_skx_npys` folder, where
-        `{name}` is the simulation name, with the file name `skx_{step}.npy`
-        where `{step}` is the simulation step (from the integrator)
-        """
-        if not os.path.exists('%s_skx_npys' % self.name):
-            os.makedirs('%s_skx_npys' % self.name)
-        name = '%s_skx_npys/m_%g.npy' % (self.name, self.step)
-
-        # The _skx_number array is defined in the SimBase class in Common
-        np.save(name, self._skx_number)
+        self.VTK.write_file(step=self.step)
